requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
psycopg[binary]==3.2.10
flask-migrate==4.0.7
python-dotenv==1.0.1
alembic==1.13.2
flask-limiter==3.7.0
reportlab==4.2.0
gunicorn==22.0.0
redis==5.0.8

